chore(web): extract Cost & performance panel to /theguard/spend/optimization
Phase 2 Chunk A panel 3 of 4. Moves the "Cost & performance" (Token
guardrails) tab out of the Guard settings tabs into a dedicated route +
features/guard/spend/optimization/Panel.tsx.

GuardToggle lifted to features/guard/GuardToggle.tsx so both settings
(Enforcement) and the extracted panel share one implementation.

Settings redirect shim now covers ?tab=notifications and ?tab=guardrails.
Token-guardrails hook + toggle handler + save-toast state all move into
the panel so the settings page no longer imports them.
